+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [0.6.0] **WIP**
+## [0.6.0]
 * feat: new internal function `strToBytesArray` to convert a `string` to a n `Uint8Array` (`strings.js`)
 * refactor!: renamed internal function `utf8ArrayToStr` to `bytesArrayToStr` and changed prototype (`strings.js`)
 * fix: close stdout & stderr pipes after execution (`process.js`)